deploy/requirements.txt
scikit-learn==1.3.0
google-cloud-storage==2.10.0
joblib==1.3.2
numpy==1.26.4
scipy==1.11.4